Backend to frontend 2

+++ backend/src/main/java/com/mythovac/backend/controller/ScoreController.java
@@ -12,6 +12,7 @@ import com.mythovac.backend.service.impl.ScoreServiceImpl;
 import com.mythovac.backend.service.impl.ScoresCommentServiceImpl;
 import com.mythovac.backend.service.impl.ScoresReleaseServiceImpl;
 import com.mythovac.backend.service.impl.UserServiceImpl;
+import jakarta.servlet.http.HttpSession;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -31,61 +32,72 @@ public class ScoreController {
         this.scoresReleaseService = scoresReleaseService;
     }
 
+    @GetMapping("get-all-scores")
+    public Result getAllScores() {
+        List<Score> res = scoreService.getAllScores();
+        return Result.success(res);
+    }
+
+    @GetMapping("get-all-avail-scores")
+    public Result getAllAvailScores() {
+        List<Score> res = scoreService.getAllAvailScores();
+        return Result.success(res);
+    }
+
     @GetMapping("/get-score-by-sid/{sid}")
     public Result getScore(@PathVariable Long sid) {
         Score res = scoreService.getScoreBySid(sid);
-        if(res == null) return Result.error("评分不存在");
-        return Result.success(res);
+        return res == null ? Result.error("评分不存在") : Result.success(res);
     }
 
     @GetMapping("/get-comments-by-sid/{sid}")
     public Result getCommentsBySid(@PathVariable Long sid) {
         List<ScoresComment> res = scoresCommentService.getScoresCommentBySid(sid);
-        if(res == null) return Result.error("评分不存在");
         return Result.success(res);
     }
 
     @GetMapping("/get-score-by-tag/{tag}")
     public Result getScoreByTag(@PathVariable String tag) {
         List<Score> res = scoreService.getScoresByTag(tag);
-        if(res == null) return Result.error("评分不存在");
         return Result.success(res);
     }
 
     @PostMapping("/add-score")
     public Result addScore(@RequestBody Score score) {
-        if(score == null || score.getTag() == null || score.getGoal() == null) {
+        if (score == null || score.getTag() == null || score.getGoal() == null) {
             return Result.error("评分内容不符");
         }
+        score.setStatus(0L);
         scoreService.insertScore(score);
         return Result.success();
     }
 
 
     @PostMapping("/add-comment")
-    public Result addComment(@RequestBody ScoresComment scoresComment) {
-        if(scoresComment == null || scoresComment.getSid() == null) {
+    public Result addComment(@RequestBody ScoresComment scoresComment, HttpSession session) {
+        Result sessionCheck = UserController.checkSession(session);
+        if (sessionCheck != null) return sessionCheck;
+
+        if (scoresComment == null || scoresComment.getSid() == null) {
             return Result.error("评分不存在");
         }
-        if(scoresComment.getUid() == null) {
-            return Result.error("用户不存在");
-        }
-        if(userService.getUserById(scoresComment.getUid()) == null){
-            return Result.error("用户不存在");
-        }
-        if(scoreService.getScoreBySid(scoresComment.getSid()) == null){
+        if (scoreService.getScoreBySid(scoresComment.getSid()) == null) {
             return Result.error("评分不存在");
         }
+        scoresComment.setUid((Long) session.getAttribute("uid"));
         scoresCommentService.insertScoresComment(scoresComment);
         return Result.success();
     }
 
     @PostMapping("/update-score")
-    public Result updateScore(@RequestBody Score score) {
-        if(score == null || score.getSid() == null) {
+    public Result updateScore(@RequestBody Score score, HttpSession session) {
+        Result sessionCheck = UserController.checkSession(session);
+        if (sessionCheck != null) return sessionCheck;
+
+        if (score == null || score.getSid() == null) {
             return Result.error("评分不存在");
         }
-        if(scoreService.getScoreBySid(score.getSid()) == null){
+        if (scoreService.getScoreBySid(score.getSid()) == null) {
             return Result.error("评分不存在");
         }
         scoreService.updateScore(score);
@@ -93,14 +105,18 @@ public class ScoreController {
     }
 
     @DeleteMapping("/delete-score-by-sid/{sid}")
-    public Result deleteScore(@PathVariable Long sid) {
-        if(scoreService.getScoreBySid(sid) == null){
+    public Result deleteScore(@PathVariable Long sid, HttpSession session) {
+        Result sessionCheck = UserController.checkSession(session);
+        if (sessionCheck != null) return sessionCheck;
+        if ((Integer) session.getAttribute("permission") != 3)
+            return Result.error("权限不足");
+
+        if (scoreService.getScoreBySid(sid) == null) {
             return Result.error("评分不存在");
         }
         scoresCommentService.deleteScoresCommentBySid(sid);
         scoreService.deleteScoreBySid(sid);
         return Result.success();
     }
-
 
 }
